# lib/api-zod/src/generated/types/quizSubmissionResult.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification for SlideShare - real-time presentation sharing
 * OpenAPI spec version: 0.1.0
 */
import type { QuizReviewItem } from './quizReviewItem';

export interface QuizSubmissionResult {
  id: string;
  quizId: string;
  studentName: string;
  score: number;
  totalQuestions: number;
  createdAt: Date;
  /**
     * @minItems 20
     * @maxItems 20
     */
  review: QuizReviewItem[];
}